test: test these commands
